requirements.txt
cryptography>=43.0.0
dnspython>=2.6.1
requests-cache>=1.2.1
lxml>=5.4.0